requirements.txt
# Web 自动化测试
selenium==4.15.0
webdriver-manager==4.0.1

# HTTP 客户端和 WebSocket
requests==2.31.0
websocket-client==1.6.4

# 测试框架和断言
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-timeout==2.2.0

# 数据处理和配置
pyyaml==6.0.1
python-dotenv==1.0.0

# 日志和报告
allure-pytest==2.13.2
colorlog==6.7.0

# Docker 客户端
docker==6.1.3

# 时间处理
python-dateutil==2.8.2

# 系统监控
psutil==5.9.6